community-profiles: Make community-profiles more modular
The package community-files contains the configuration-definitions of several
wireless communities (and Freifunk) over the world. This commit removes the
unnecessary dependency on freifunk-common. Thus we can use them more independently.

For example: Freifunk Falter, a project of several Freifunk-Berlin-Developers, uses
its own freifunk-common package (falter-common). As this package contains only plain
text-files, ther is not rally a tru dependency on the freifunk-common package.

Signed-off-by: Martin Hübner <martin.hubner@web.de>